requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.7.0
Django==4.1.5
fonttools==4.38.0
html5lib==1.1
Pillow==9.4.0
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.5.0
pyphen==0.13.2
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
tomli==2.0.1
tzdata==2022.7
weasyprint==57.2
webencodings==0.5.1
zopfli==0.2.2
pycups==2.0.1